cert_agent/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-core>=2.0.0
openai>=1.0.0
python-multipart>=0.0.6
httpx>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: For better performance
orjson>=3.9.0